+++ src/upper_bound.cc
@@ -4,6 +4,6 @@ upper_bound::upper_bound () { }
 
 upper_bound::~upper_bound () { }
 
-int upper_bound::compute_simple_upper_bound (node* t1, node* t2) {
+double upper_bound::compute_simple_upper_bound (node* t1, node* t2) {
 	return (t1->get_subtree_size() + t2->get_subtree_size());
 }
